Catalog fetching refactor
