+++ addon/components/cardflow.hbs
@@ -33,7 +33,7 @@
         @mode="view"
         @format="table"
         @class="cardflow__isolated-collection"
-        @model={{this.isolatedCollection}}
+        @model={{@isolatedCatalog}}
         @transition={{this.displayCatalogItem}}
         @sort={{this.sort}}
         @search={{this.search}}
@@ -99,7 +99,7 @@
             </section>
             <section class="cardflow-card__section">
               <div class="field-renderer__title">Catalog</div>
-              <Cards::MasterCollection @model={{this.catalog}} @transition={{fn this.displayCatalog this.catalog.id}} />
+              <Cards::MasterCollection @model={{@catalog}} @transition={{fn this.displayCatalog @catalog.id}} />
             </section>
             <section class="cardflow-card__section cardflow-proposal-card__effective-date">
                 Effective Date: <strong>September 1, 2020</strong>
