+++ tests/dummy/app/controllers/media-registry/cardflow.js
@@ -1,0 +1,91 @@
+import Controller from '@ember/controller';
+import { tracked } from '@glimmer/tracking';
+import { action } from '@ember/object';
+import { dasherize } from '@ember/string';
+import { fetchCollection } from 'dummy/media';
+
+export default class MediaRegistryCardflowController extends Controller {
+  @tracked isolatedCollection = this.getIsolatedCollection(this.catalog.id);
+
+  catalog = {
+    id: 'batch-f',
+    type: 'catalog',
+    title: 'Batch F',
+    catalog_title: 'Batch F',
+    catalog_description: 'Transfer to CRD Records',
+    number_of_songs: 16,
+    selected_art: [
+      "media-registry/covers/thumb/Sunlight.jpg",
+      "media-registry/covers/thumb/Change-Is-Good.jpg",
+      "media-registry/covers/thumb/Full-Moon.jpg",
+      "media-registry/covers/thumb/Love-Never-Dies.jpg",
+      "media-registry/covers/thumb/Animals.jpg"
+    ]
+  }
+
+  @action
+  async getIsolatedCollection(id) {
+    const data = await fetchCollection('all_tracks_combined');
+
+    let items = data.filter(item => {
+      if (item.catalog) {
+        return item.catalog.map(catalog => {
+          let catalogId = dasherize(catalog.trim());
+          return catalogId === id;
+        }).includes(true);
+      }
+    });
+
+    this.isolatedCollection = {
+      title: id,
+      type: 'collection',
+      collection: items,
+      itemType: 'masters',
+      columns: [
+        {
+          name: 'Title',
+          valuePath: 'song_title',
+          isFixed: 'left',
+          width: 350,
+        },
+        {
+          name: 'Artist',
+          valuePath: 'artist',
+          width: 250,
+        },
+        {
+          name: 'Release Title',
+          valuePath: 'album',
+          width: 250,
+        },
+        {
+          name: 'Artwork',
+          valuePath: 'cover_art',
+          width: 175,
+          isSortable: false,
+        },
+        {
+          name: 'Release Type',
+          valuePath: 'type_of_album',
+          width: 250,
+        },
+        {
+          name: 'Genre',
+          valuePath: 'genre',
+          width: 250,
+        },
+        {
+          name: 'Length',
+          valuePath: 'length',
+          width: 250,
+          sortType: 'numeric'
+        },
+        {
+          width: 0,
+          isFixed: 'right',
+          isSortable: false
+        },
+      ],
+    };
+  }
+}
