+++ tests/dummy/app/styles/app.css
@@ -4,7 +4,6 @@
 @import 'media-registry/item.css';
 @import 'media-registry/item-edit.css';
 @import 'media-registry/agreements.css';
-@import 'components/sidebar-related-card.css';
 
 
 html, body {
