+++ tests/dummy/app/templates/media-registry/cardflow.hbs
@@ -1,1 +1,5 @@
-<Cardflow @model={{@model}} />
+<Cardflow
+  @model={{@model}}
+  @isolatedCatalog={{this.isolatedCollection}}
+  @catalog={{this.catalog}}
+/>
